+++ src/modules/access-wallet/hardware/handlers/hardwares/bitbox02/index.js
@@ -136,6 +136,7 @@ class BitBox02Wallet {
 const createWallet = async () => {
   const _bb02Wallet = new BitBox02Wallet();
   await _bb02Wallet.connect();
+  _bb02Wallet.errorHandler = errorHandler;
   return _bb02Wallet;
 };
 createWallet.errorHandler = errorHandler;